+++ Assets/Game/Game.cs
@@ -10,10 +10,6 @@ public class Game : MonoBehaviour {
     private GameObject controllerSettings;
     private EndGame endGame;
 
-    //Touch input (swipe)
-    private Vector2 touchStartPos = new Vector2(0, 0);
-    private Vector2 touchEndPos = new Vector2(0, 0);
-
     void Awake() {
         points = FindObjectOfType<pointsCounter>();
         managerBlocks = FindObjectOfType<blocksManager>();
@@ -22,23 +18,6 @@ public class Game : MonoBehaviour {
         controllerSettings = GameObject.FindGameObjectWithTag("Settings");
         endGame = FindObjectOfType<EndGame>();
         gameObject.SetActive(false);
-    }
-
-    void Update() {
-        //Input
-        switch (Settings.selectedInput) {
-            case Settings.InputMode.BUTTONS:
-                //selectedInputButtons();
-                break;
-
-            case Settings.InputMode.TOUCH_TAP:
-                //selectedInputTap();
-                break;
-
-            case Settings.InputMode.TOUCH_SWIPE:
-                selectedInputSwipe();
-                break;
-        }       
     }
 
 	public void newGame() {
@@ -51,25 +30,4 @@ public class Game : MonoBehaviour {
     }
 
     public int getPoints() { return points.getPoints(); }
-
-    private void selectedInputSwipe() {
-        if (Input.touchCount > 0) {
-            if (!GamePause.isPaused() && !controllerSettings.activeInHierarchy && !endGame.gameObject.activeInHierarchy) {
-                if (Input.GetTouch(0).phase == TouchPhase.Began) touchStartPos = Input.GetTouch(0).position;
-                else if (Input.GetTouch(0).phase == TouchPhase.Ended) {
-                    touchEndPos = Input.GetTouch(0).position;
-
-                    if (Mathf.Abs(touchEndPos.x - touchStartPos.x) > Mathf.Abs(touchEndPos.y - touchStartPos.y)) {
-                        if (touchEndPos.x > touchStartPos.x) managerBlocks.getBlock().GetComponent<Block>().turnRight();
-                        else managerBlocks.getBlock().GetComponent<Block>().turnLeft();
-                    }
-                    else {
-                        if (touchEndPos.y > touchStartPos.y) managerBlocks.getBlock().GetComponent<Block>().rotate();
-                        else managerBlocks.getBlock().GetComponent<Block>().speedUp = true;
-                    }
-                }
-            }
-            else if (GamePause.isPaused() && !controllerSettings.activeInHierarchy) GamePause.deactivate();
-        }
-    }
 }
